backend/requirements.txt
fastapi
uvicorn
redis
celery
python-dotenv
pydantic
